0.1.1
This commit updates the version number to 0.1.1 and adds patient data fetching functionality to CustomAppBar using http package and json parsing. PatientData class is used to store the name and username of the patient fetched from the API. The isLoading variable is used to show a CircularProgressIndicator while the data is being fetched. The FocusedMenuItem with 'Haidaruddin Muhammad Ramdhan' and '1301204459' as its title is replaced with the actual name and username fetched from the API. Additionally, the code has been formatted and several unused imports have been removed from the file 'lib/custom_app_bar.dart'.