0.1.5
This commit updates the version in `pubspec.yaml` to `0.1.5`. Additionally, it refactors the `KonsultasiScreen` widget in `lib/pasien/konsultasi_pasien.dart`. The constructor is now formatted with line breaks for better readability. The `_loadItems()` method also has a reduced delay time of one second instead of two. Lastly, the `ListView` in the widget now has a vertical scroll direction and a disabled scrolling physics. Each `ListTile` item is now wrapped in a `Padding` widget with bottom padding to add spacing between the list items. The `ListTile` now has its content padding, dense and tile color properties set for better visual appearance. A trailing navigation icon and an `onTap` callback were added to the `ListTile` as well, although the callback function is currently empty.