0.1.6
The previous body container did not have a fixed size, causing the layout to be unpredictable. The new implementation uses SizedBox to set the width and height of the body container. Additionally, the child of the body container is changed to SingleChildScrollView, which allows for scrolling if the content exceeds the container size. The layout is updated with a search bar and a list of cards that display dummy data. Also, update the version in pubspec.yaml to 0.1.6.