0.2.1
Refactor _ReservasiScreenState widget to improve its UI and user experience by changing the Container to a SizedBox and adding SingleChildScrollView to enable scrolling of the Column containing all its child widgets. The ListView.builder is also added to display a list of dummy data. In addition, the pubspec.yaml file is updated to version 0.2.1.