0.2.3
This commit adds a new method to show a dialog box that prompts the user to confirm whether they want to accept or reject a reservation. The dialog is shown when the corresponding button is pressed and it displays a message asking the user if they are sure they want to perform the action. Additionally, the version number of the project has been updated from 0.2.2 to 0.2.3 in the pubspec.yaml file.